Cleaned up Zombies with HatZombie class
Added a HatZombie abstract class to make implementing modified basic zombies easier.
